Update Temp and LCD
I added a blue and red LED to tell the user when it is above or below 75 degrees. I'm trying to get them to blink when it is above or below a certain threshold, but I keep getting stuck in while loops. I'll be doing more research to figure out how to fix this situation.